fix: index name changed

+++ api/src/entities/customer_wise_day_wise_redemption_summary_mv.ts
@@ -10,11 +10,11 @@ import { ViewEntity, ViewColumn, Index } from 'typeorm';
       c.external_id AS customer_external_id,
       r.redemption_date AS date,
 
-      COUNT(r.redemption_id)::numeric AS total_redemptions,
+      COUNT(r.redemption_id) AS total_redemptions,
 
-      SUM(r.base_order_value)::numeric AS gross_sale,
-      SUM(r.discount)::numeric AS total_discount,
-      SUM(r.base_order_value - r.discount)::numeric AS net_sale,
+      SUM(r.base_order_value) AS gross_sale,
+      SUM(r.discount) AS total_discount,
+      SUM(r.base_order_value - r.discount) AS net_sale,
 
       NOW() AS created_at,
       NOW() AS updated_at
